pyproject.toml
[project]
name = "tf2"
version = "2.0.0"
description = "TalentFlow 2 Project"
authors = [
    {name = "Shun-Fu Hu", email = "husohome98@gmail.com"},
]
requires-python = ">=3.10"
readme = "README.md"
license = {text = "MIT"}
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "python-multipart>=0.0.6",
    "pydantic>=2.6.0",
    "langchain>=0.1.0",
    "pypdf>=4.0.0",
    "python-magic>=0.4.27"
]

[project.urls]
Repository = "https://github.com/husohome/tf2"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["tf2"]
